+++ packages/design-system-react/src/components/IconAlert/index.ts
@@ -1,0 +1,3 @@
+export { IconAlertSeverity } from '@metamask/design-system-shared';
+export { IconAlert } from './IconAlert';
+export type { IconAlertProps } from './IconAlert.types';