+++ packages/design-system-shared/src/types/KeyValueSelect/index.ts
@@ -1,0 +1,5 @@
+export {
+  KeyValueSelectVariant,
+  type KeyValueSelectPropsShared,
+  type KeyValueSelectSelectButtonPropsShared,
+} from './KeyValueSelect.types';